# Metal: Localized logic in `ggml_metal_graph_compute`, minor performance improvement