This PR combines the 3 loops in the `ggml_metal_graph_compute` function into one. The changes can produce a (very) minor performance increase on occasion, but the main focus was to try and simplify the structure of the function a bit and reduce reliance on external references.

- `command_buffers` and `command_encoders` are removed from the context, as they were only used locally in the compute function. However this meant both having to worry about global state in the context but also required the runtime to jump to the heap from/to inside the compute function.
- Each command buffer is retrieved on each loop of the n_cb count. The method is changed from `commandBuffer` to `commandBufferWithUnretainedReferences` since the buffers it accesses are inside the `ctx` and managed as part of its lifecycle, so there is no need for an additional retention by the command buffer during the compute pass. The reference is managed locally, hopefully letting the compiler put it on the stack for performance.
- The `encoder` used for each command buffer is requested and accessed inside the `dispatch` block, keeping the reference thread-local, and obviating the need for maintaining the `command_encoders` array in the context.
- `GGML_METAL_MAX_COMMAND_BUFFERS` is not needed, as there is no longer need for a fixed array to track the command buffers either.
- The completion and status check for each buffer is handled inside the dispatch thread that encodes the buffer, so that all the status checks can also be performed in parallel, same as the encoder submission.
- `all_buffers_succeeded` has a theoretical thread race condition where multiple buffers fail together and `false` is written to it concurrently - however in practise concurrent writes of `false` can't cause corruption, and no other writes are performed.

All tests seem to pass with these changes and my limited real-world testing, but I would definitely appreciate more sanity checks from more experienced contributors :)

#### Performance

The change is minor. This is a representative comparison of before/after on a real-world run.

```
./bin/main -m dolphin-2.6-mixtral-8x7b.Q5_K_M.gguf -c 0 --temp 0.2 --repeat_penalty 1.0 -n -1 -p "<|user|>\nWrite long essay explaining how victorian romantic poets directly contributed to the evolution of the american sitcom<|endoftext|>\n<|assistant|>\n" -t 16 -s 123
```

```
llama_print_timings:        load time =    1515.67 ms
llama_print_timings:      sample time =      11.66 ms /   589 runs   (    0.02 ms per token, 50527.58 tokens per second)
llama_print_timings: prompt eval time =     529.39 ms /    43 tokens (   12.31 ms per token,    81.22 tokens per second)
llama_print_timings:        eval time =   15534.36 ms /   588 runs   (   26.42 ms per token,    37.85 tokens per second)
llama_print_timings:       total time =   16135.30 ms /   631 tokens
```

```
llama_print_timings:        load time =    1513.06 ms
llama_print_timings:      sample time =      11.48 ms /   589 runs   (    0.02 ms per token, 51324.50 tokens per second)
llama_print_timings: prompt eval time =     528.20 ms /    43 tokens (   12.28 ms per token,    81.41 tokens per second)
llama_print_timings:        eval time =   15463.36 ms /   588 runs   (   26.30 ms per token,    38.03 tokens per second)
llama_print_timings:       total time =   16061.03 ms /   631 tokens
```